% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_call.R
\name{act_call}
\alias{act_call}
\title{Return a part of a data frame corresponding to a condition}
\usage{
act_call(data, unique_table, scenario_nb, rt = "df")
}
\arguments{
\item{data}{data frame}

\item{unique_table}{List of unique conditions, can be created with AOB::unique_table.}

\item{scenario_nb}{Row of unique_table to compute, compute every row if not specified. Useful in loop.}

\item{rt}{return format}
}
\value{
Data frame with asked condition.
}
\description{
Return a part of a data frame corresponding to a condition
}
\examples{
data("CO2")
ut <- unique(CO2[, 1:3])
uptake_mean <- data.frame()
for (i in 1:nrow(ut)) {uptake_mean <- rbind(uptake_mean, Rmisc::group.CI(uptake ~ Plant, act_call(CO2, ut, i)))}
}
