% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_mfrow.R
\name{act_mfrow}
\alias{act_mfrow}
\title{Calculate the size of the graphic windows base on the data set.}
\usage{
act_mfrow(df, col = "all")
}
\arguments{
\item{df}{data frame}

\item{col}{column of the parameters analysed}
}
\value{
c(nrow,ncol)
}
\description{
Calculate the size of the graphic windows base on the data set.
}
