% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anotations_all.R
\name{annotations_all}
\alias{annotations_all}
\title{All annotations in a corpus}
\usage{
annotations_all(x)
}
\arguments{
\item{x}{Corpus object.}
}
\value{
Data.frame.
}
\description{
Merges annotations from all transcripts in a corpus and returns a data frame.
}
\examples{

library(act)

#Get data frame with all annotations
allannotations <- act::annotations_all(examplecorpus)

#Have a look at the number of annotations
nrow(allannotations)

}
