% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations_insert_from_search_to_tier.R
\name{annotations_insert_from_search_to_tier}
\alias{annotations_insert_from_search_to_tier}
\title{Inserts search results as annotations in tiers}
\usage{
annotations_insert_from_search_to_tier(
  x,
  s,
  destTier = "destinationTier",
  destTierAddMissing = TRUE,
  contentFromColname = "hit",
  interruptOverlap = FALSE,
  filterTranscriptNames = NULL,
  filterTierNames = NULL,
  collapseString = " | "
)
}
\arguments{
\item{x}{Corpus object.}

\item{s}{Search object.}

\item{destTier}{Character string; name of the tier to which the hit should be copied (if no copying is intended set to NA).}

\item{destTierAddMissing}{Logical; if \code{TRUE} the destination tier will be added if missing in the transcript object, if \code{FALSE} an error will be raised if the destination tier is missing.}

\item{contentFromColname}{Character string; names of the search results column from which the content of the new annotation shall be copied. Chose for example "resultID", “hit” or “content” or the name of any column that you have added to the results data frame.}

\item{interruptOverlap}{Integer; How to proceed in case of overlapping search results: TRUE=insertion will be interrupted, FALSE=overlapping annotations will be merged.}

\item{filterTranscriptNames}{Vector of character strings; names of the transcripts to be included from search results.}

\item{filterTierNames}{Character string; names of the tiers to be included form search results.}

\item{collapseString}{Character string; will be used to collapse multiple search results into one string.}
}
\value{
Corpus object.
}
\description{
The function will insert the results of a search as annotations into a specified destination tier.
Results from different tiers will all be inserted into the same destination tier.
It can be specified how overlapping search results or search results from the same annotation will be handled.
Please note that hits with exactly the same start and end time will always be merged (e.g. they are not treated as overlapping).
}
\examples{
library(act)

# Have a look at the first transcript in the example corpus:
printtranscript <- act::export_txt(examplecorpus@transcripts[[1]])
cat(printtranscript)
# In line 01 there is the word "UN".

# Replace this word by "XXX" in the entire corpus
test <- act::annotations_replace_copy(x=examplecorpus,
									  pattern="\\\\bUN\\\\b",
									  replacement="XXX")

# Have a look at the first transcript in the corpus object test:
printtranscript <- act::export_txt(test@transcripts[[1]])
cat(printtranscript)
# In line 01 there is now "XXX" instead of "UN"

# Insert a tier called "newTier" into all transcripts in the corpus:
for (t in examplecorpus@transcripts) {
	sortVector <- c(t@tiers$name, "newTier")
	examplecorpus <- act::tiers_sort(x=examplecorpus,
	sortVector=sortVector,
	filterTranscriptNames=t@name,
	tiersAddMissing=TRUE)
}
# Check that the first transcript now contains the newTier
examplecorpus@transcripts[[1]]@tiers

# Now replace "UN" by "YYY" in the entire corpus and
# copy the search hit to "newTier".
test <- act::annotations_replace_copy(x=examplecorpus,
									  pattern="\\\\bUN\\\\b",
									  replacement="YYY",
									  destTier = "newTier")

# Have a look again at the first transcript in the corpus object test.
printtranscript <- act::export_txt(test@transcripts[[1]])
cat(printtranscript)
# In line 01 you see that "UN" has been replaced by "YYY.
# In line 02 you see that it has been copied to the tier "newTier".

# If you only want to copy a search hit but not replace it in the original
# leave replacement="", which is the default
test <- act::annotations_replace_copy(x=examplecorpus,
									  pattern="\\\\bUN\\\\b",
									  destTier = "newTier")
printtranscript <- act::export_txt(test@transcripts[[1]])
cat(printtranscript)
# In line 01 you see that "UN" has been maintained.
# In line 02 you see that "UN" it has been copied to the tier "newTier".
}
