% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus__class.R
\name{corpus_alldata}
\alias{corpus_alldata}
\title{Get all data in transcripts}
\usage{
corpus_alldata(x)
}
\arguments{
\item{x}{Corpus object.}
}
\value{
Data.frame.
}
\description{
Merges data from all transcripts in a corpus and returns a data frame.
}
\examples{

library(act)

#Get data frame with all data
alldata <- act::corpus_alldata(examplecorpus)

#Have a look at the number of annotations
nrow(alldata)

}
