% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus__class.R
\name{corpus_overview}
\alias{corpus_overview}
\title{Information about all transcripts in a corpus object}
\usage{
corpus_overview(x)
}
\arguments{
\item{x}{Corpus object.}
}
\value{
Data frame.
}
\description{
Information about all transcripts in a corpus object
}
\examples{
library(act)

myoverview<- act::corpus_overview(examplecorpus)
myoverview

}
