% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_examplecorpus.R
\docType{data}
\name{examplecorpus}
\alias{examplecorpus}
\title{Example corpus for the act package}
\format{
An object of class \code{"corpus"}
}
\source{
\itemize{
\item GAT: Ehmer, Oliver/Satti, Luis Ignacio/Martinez, Angelita/Pfaender, Stefan (2019): Un sistema para transcribir el habla en la interaccion: GAT 2.0 Gespraechsforschung - Online-Zeitschrift zur verbalen Interaktion (www.gespraechsforschung-ozs.de) 20, 64-114. http://www.gespraechsforschung-online.de/2019.html
\item SYNC:	Ehmer, Oliver (2020, in press): Synchronization in demonstrations. Multimodal practices for instructing body knowledge. Linguistics Vanguard. https://www.degruyter.com/view/journals/lingvan/lingvan-overview.xml
}
}
\usage{
data(examplecorpus)
}
\description{
Example corpus with data loaded from the example annotations files that come with the package
}
\details{
You can download the corresponding media files from
\itemize{
\item www.oliverehmer.de in the section "Digital Humanities or
\item from Github: https://github.com/oliverehmer/act_examplecorpus/.
Alternatively you can use the download commands in the example section below.
}
}
\examples{
library(act)

# Summary of the data in the corpus
examplecorpus

# Summary of the data in th second transcripts in the corpus
examplecorpus@transcripts[[2]]

\dontrun{
# Download example corpus with media files
destinationpath <- 
temp <- tempfile()
download.file(options()$act.examplecorpusURL, temp)
unzip(zipfile=temp, exdir=destinationpath)

# Set the URL for the ZIP archive
url <- options()$act.examplecorpusURL

# Define output path
output_dir <- "/EXISTING_FOLDERON_YOUR_COMPUTER/examplecorpus"
output_path <- file.path(output_dir, "act_examplecorpus.zip")

# Download the ZIP file
download.file(url, output_path, mode = "wb")

# Unzip it to the output directory
unzip(output_path, exdir = output_dir)

# Rename the extracted folder
folder_old <- file.path(output_dir, "act_examplecorpus-main")
folder_new <- file.path(output_dir, "act_examplecorpus")
if (dir.exists(folder_new)) unlink(folder_new, recursive = TRUE)
file.rename(folder_old, folder_new)

}
}
\seealso{
, \link{corpus_new}, \link{corpus_import}
}
\keyword{datasets}
