% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_docx_styles_load.R
\name{export_docx_styles_load}
\alias{export_docx_styles_load}
\title{Load styles matrix for .docx transcripts}
\usage{
export_docx_styles_load(path = NULL, encoding = "UTF-8", path_docx = NA)
}
\arguments{
\item{path}{Character string; path to the export styles matrix .csv.  If argument is left open, the default export styles data frame of the package will be returned.}

\item{encoding}{Character string; encoding of the file.}

\item{path_docx}{Character string; Path to a template file. If given, it will be checked if all styles defined in exportStylesMatrix are present in the .docx file. if \code{NA} nothing will be checked.}
}
\value{
Data.frame
}
\description{
This function is only for checking how the export styles matrix (as .csv) will be loaded internally.
}
\examples{
library(act)

# An example replacement matrix comes with the package.
path <- system.file("extdata", "normalization", "normalizationMatrix.csv", package="act")

# Load the matrix
mymatrix <- act::matrix_load(path)

# Have a look at the matrix
colnames(mymatrix)
mymatrix

#the original path of the matrix is stored in the attributes
attr(mymatrix, 'path') 
}
