% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_docx_styles_save.R
\name{export_docx_styles_save}
\alias{export_docx_styles_save}
\title{Save styles matrix for .docx transcripts}
\usage{
export_docx_styles_save(exportStyles, path, encoding = "UTF-8")
}
\arguments{
\item{exportStyles}{Data frame; export styles matrix.}

\item{path}{Character string; path where the matrix will be saved.}

\item{encoding}{Character string; encoding of the file.}
}
\value{
nothing
}
\description{
Save styles matrix for .docx transcripts
}
\examples{
library(act)

# An example replacement matrix comes with the package.
path <- system.file("extdata", "normalization", "normalization_matrix.csv", package="act")

# Load the matrix
mymatrix <- act::matrix_load(path)

# ' # Create temporary file path
path <- tempfile(pattern = "mymatrix", tmpdir=tempdir(),  fileext = ".csv")

# It makes more sense, however, to you define a destination folder
# that is easier to access on your computer:
\dontrun{
path <- file.path("PATH_TO_AN_EXISTING_FOLDER_ON_YOUR_COMPUTER",
                  "mymatrix.csv")
}

# Save the matrix
act::matrix_save(mymatrix, path=path)

}
