% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_eaf.R
\name{export_eaf}
\alias{export_eaf}
\title{Export 'ELAN' .eaf file}
\usage{
export_eaf(
  t,
  pathOutput = NULL,
  filterTierNames = NULL,
  filterSectionStartsec = NULL,
  filterSectionEndsec = NULL,
  createMediaLinks = TRUE,
  mediaMaxFiles = NULL,
  mediaWriteOffset = FALSE
)
}
\arguments{
\item{t}{Transcript object; transcript to be exported.}

\item{pathOutput}{Character string; path where .eaf file will be saved.}

\item{filterTierNames}{Vector of character strings; names of tiers to be included. If left unspecified, all tiers will be exported.}

\item{filterSectionStartsec}{Double; start of selection in seconds.}

\item{filterSectionEndsec}{Double; end of selection in seconds.}

\item{createMediaLinks}{Logical; if \code{TRUE} media links will be created.}

\item{mediaMaxFiles}{Integer or \code{NULL}; maximum number of media files to link (\code{NULL} = all).}

\item{mediaWriteOffset}{Logical; if \code{TRUE} a negative \code{TIME_ORIGIN} attribute (in milliseconds) is written for clip media (rows with \code{kind == "clip"} and \code{startsec > 0}), derived from \code{startsec}. Default \code{FALSE} (no offset). Note: clip media are currently not linked at all in the exported .eaf, because 'ELAN' does not support negative media offsets - only full recordings (offset 0) are linked. This parameter takes effect once clip linking is re-enabled in the code.}
}
\value{
Contents of the .eaf file (only if \code{pathOutput} is left empty)
}
\description{
Advice: In most situations it is more convenient to use \code{act::corpus_export} for exporting annotation files.
}
\details{
The .eaf file will be written to the file specified in \code{pathOutput}.
If \code{pathOutput} is left empty, the function will return the contents of the .eaf itself.
}
\examples{
library(act)

# Get the transcript you want to export
t <- examplecorpus@transcripts[[1]]

# Create temporary file path
path <- tempfile(pattern = t@name, tmpdir = tempdir(), fileext = ".eaf")

# It makes more sense, however, to you define a destination folder
# that is easier to access on your computer
\dontrun{
path <- file.path("PATH_TO_AN_EXISTING_FOLDER_ON_YOUR_COMPUTER",
                   paste(t@name, ".eaf", sep=""))
}

# Export WITH media links
act::export_eaf(t=t, pathOutput=path)

# Export WITHOUT media links
act::export_eaf(t=t, pathOutput=path, createMediaLinks = FALSE)

}
\seealso{
\code{corpus_export}, \link{export_exb}, \link{export_txt}, \link{export_rpraat}, \link{export_srt}, \link{export_textgrid}, \link{export_docx}
}
