% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_edl.R
\name{export_edl}
\alias{export_edl}
\title{Export .edl marker file for '"'Davinci Resolve'}
\usage{
export_edl(
  t,
  pathOutput = NULL,
  filterTierNames = NULL,
  filterSectionStartsec = NULL,
  filterSectionEndsec = NULL,
  fps = 50
)
}
\arguments{
\item{t}{Transcript object; transcript to be saved.}

\item{pathOutput}{Character string; path where .edl will be saved.}

\item{filterTierNames}{Vector of character strings; names of tiers to be included. If left unspecified, all tiers will be exported.}

\item{filterSectionStartsec}{Double; start of selection in seconds.}

\item{filterSectionEndsec}{Double; end of selection in seconds.}

\item{fps}{Double; Frame rate per seconds of your project, e.g. 60, 50, 30, 29}
}
\value{
Contents of the .edl file (only if \code{pathOutput} is left empty)
}
\description{
Advice: In most situations it is more convenient to use \code{act::corpus_export} for exporting annotation files.
}
\details{
Creates a 'event list text' .edl file with definitions for Blackmagic DaVinci Resolve
It will be written to the file specified in \code{pathOutput}.
If \code{pathOutput} is left empty, the function will return the contents of the .edl itself.
}
\examples{
library(act)

# Get the transcript you want to export
t <- examplecorpus@transcripts[[1]]

# Create temporary file path
path <- tempfile(pattern = t@name, tmpdir = tempdir(),
                 fileext = ".srt")

# It makes more sense, however, to you define a destination folder
# that is easier to access on your computer:
\dontrun{
path <- file.path("PATH_TO_AN_EXISTING_FOLDER_ON_YOUR_COMPUTER",
                    paste(t@name, ".srt", sep=""))
}

# Export
act::export_srt(t=t, pathOutput=path)

}
\seealso{
\code{corpus_export}, \link{export_eaf}, \link{export_exb}, \link{export_srt}, \link{export_txt}, \link{export_docx}, \link{export_rpraat}, \link{export_textgrid}
}
