% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_print.R
\name{ggplot_print}
\alias{ggplot_print}
\title{Multiple plot with ggplot2
use "mfrow" from par() with ggplot}
\usage{
ggplot_print(plot)
}
\arguments{
\item{plot}{ggplot input}
}
\description{
Multiple plot with ggplot2
use "mfrow" from par() with ggplot
}
