% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\title{Plot following Annals of Botany rules}
\usage{
graph(
  x,
  y,
  z,
  data,
  LoopSize = 3,
  mfrow = NULL,
  i = 1,
  type = "p",
  palette = 2,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  xlim = TRUE,
  ylim = TRUE,
  outer = TRUE,
  ann = TRUE,
  lg.lab = "",
  lg.pos = "bottomright",
  lg.nb = c(1:1000),
  Mean = FALSE,
  Interval = FALSE,
  ErrorBar = FALSE,
  reg = FALSE,
  reg.itc = NULL,
  reg_pos = "left",
  reg_pos2 = "right",
  hist = FALSE,
  x2 = NULL,
  y2 = NULL,
  xlim2 = TRUE,
  ylim2 = TRUE,
  path = NULL
)
}
\arguments{
\item{x}{character. set of data to draw}

\item{y}{character. set of data to draw}

\item{z}{character. set of data to draw}

\item{data}{data frame. First three columns must be the factors parameters.}

\item{LoopSize}{number of factor}

\item{mfrow}{plot number by graphical window. c(ncol, nrow)}

\item{i}{numeric. if i > 1 changes of the graphical window are ignored. Can have decimal to adjust plot letters if needed.}

\item{type}{what type of plot should be drawn. Default is points}

\item{palette}{1 = black and white palette, 2 = color palette. Other palette are accepted if fully detailed : palette(new.palette)}

\item{title}{plot title}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{xlim}{Plot limit. Either c(min, max), NULL or TRUE. If TRUE all the plot will have the same value, if NULL each plot have their own value.}

\item{ylim}{Plot limit. Either c(min, max), NULL or TRUE. If TRUE all the plot will have the same value, if NULL each plot have their own value.}

\item{outer}{logical. If TRUE only one xlab and ylab are writen for all the plot. If FALSE, each plot get the xlab and ylab.}

\item{ann}{logical. if TRUE annotation are displayed}

\item{lg.lab}{a character or expression vector of length ≥ 1 to appear in the legend.}

\item{lg.pos}{coordinate to position the legend}

\item{lg.nb}{numeric. plot number where the legend should be displayed. Default is on all plots.}

\item{Mean}{logical. If TRUE the mean of the y is used instead of raw value.}

\item{Interval}{logical. If TRUE display the confidence interval of the y mean}

\item{ErrorBar}{logical. If TRUE display the error bar of the y mean}

\item{reg}{Draw the linear regression of the plot. If "nls" or "lm" return a second order regression. If numeric, return a regression from the specified order, using the lm function.}

\item{reg.itc}{numeric. Force the intercept of the regression}

\item{reg_pos}{regression position on the plot}

\item{reg_pos2}{regression position on the plot}

\item{hist}{logical. If TRUE display the distribution of the data.}

\item{x2}{second set of data to draw on the same plot}

\item{y2}{second set of data to draw on the same plot}

\item{xlim2}{plot parameter of the second data set}

\item{ylim2}{plot parameter of the second data set}

\item{path}{Path where the plot are saved as png. If empty open a new window instead.}
}
\value{
a list of plot
}
\description{
Plot following Annals of Botany rules
}
\examples{
data(CO2)
graph("conc", "uptake", "Type", CO2, 1, Mean = TRUE, ErrorBar = TRUE)
}
