% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_cutlist_save.R
\name{helper_cutlist_save}
\alias{helper_cutlist_save}
\title{Helper: Save cutlist}
\usage{
helper_cutlist_save(
  cutlistMac = NULL,
  cutlistWin = NULL,
  outFolder,
  outFilename
)
}
\arguments{
\item{cutlistMac}{Character string; Content if file, if \code{NULL} no file will be saved.}

\item{cutlistWin}{Character string; Content if file, if \code{NULL} no file will be saved.}

\item{outFolder}{Character string; Destination folder.}

\item{outFilename}{Character string; Destination filename}
}
\description{
Saves FFMPEG cut list fpr mac or windows
For windows: simly saves to a .cmd file
For mac: saves as a shell script and makes it executable (also adds "#!/bin/sh")
}
\examples{
library(act)

# --- Create two tier tables from scratch
tierTable1 <- act::helper_tiers_new_table(c("a","b","c","d"),
c("IntervalTier", "TextTier","IntervalTier","TextTier"))

tierTable2 <- act::helper_tiers_new_table(c("a","b","x","y"),
c("IntervalTier", "TextTier","IntervalTier","TextTier"))

tierTable3 <- act::helper_tiers_merge_tables(tierTable1,tierTable2)
tierTable3

}
