% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_ffmpeg_path.R
\name{helper_ffmpeg_path}
\alias{helper_ffmpeg_path}
\title{Helper: Resolve the path to the FFmpeg executable}
\usage{
helper_ffmpeg_path()
}
\value{
Character string; path to the FFmpeg executable or \code{"ffmpeg"}.
}
\description{
Returns the path to the FFmpeg executable using the following cascade:
\enumerate{
\item Option \code{act.path.ffmpeg}, if it is set to a concrete path (any value other than the default \code{"ffmpeg"}).
\item \code{Sys.which("ffmpeg")}, if FFmpeg is found in the PATH of the R session.
\item Known installation locations (\code{/opt/homebrew/bin}, \code{/usr/local/bin}, \code{/usr/bin}).
\item The literal \code{"ffmpeg"} as last resort (resolution is left to the shell).
}
}
\examples{
act::helper_ffmpeg_path()

}
\seealso{
\link{helper_ffprobe_path}
}
