% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_history_changed.R
\name{helper_history_mark_export}
\alias{helper_history_mark_export}
\title{Mark transcripts as exported}
\usage{
helper_history_mark_export(
  x,
  transcriptNames = NULL,
  formats = NULL,
  folderOutput = NULL
)
}
\arguments{
\item{x}{Corpus object.}

\item{transcriptNames}{Vector of character strings; names of the exported
transcripts. If left unspecified, all transcripts are marked.}

\item{formats}{Vector of character strings; the exported formats.}

\item{folderOutput}{Character string; the output folder.}
}
\value{
Corpus object with the updated histories.
}
\description{
Appends an \code{export} entry to the \code{@history} of the given
transcripts of a corpus. \code{\link{corpus_export}} does not change the
corpus it is given; call this helper afterwards so that
\code{\link{helper_history_changed}} knows that the current state has been
written to disk.
}
\seealso{
\link{helper_history_changed}, \link{corpus_export}
}
