% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_progress.R
\name{helper_progress_tick}
\alias{helper_progress_tick}
\title{Helper: Advance progress bar by one tick}
\usage{
helper_progress_tick()
}
\description{
Helper: Advance progress bar by one tick
}
\examples{
library(act)

# --- Create two tier tables from scratch
tierTable1 <- act::helper_tiers_new_table(c("a","b","c","d"),
c("IntervalTier", "TextTier","IntervalTier","TextTier"))

tierTable2 <- act::helper_tiers_new_table(c("a","b","x","y"),
c("IntervalTier", "TextTier","IntervalTier","TextTier"))

tierTable3 <- act::helper_tiers_merge_tables(tierTable1,tierTable2)
tierTable3

}
