% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_profile.R
\name{helper_style_list}
\alias{helper_style_list}
\title{Helper: Style profiles that can be found}
\usage{
helper_style_list()
}
\value{
Named vector of character strings; paths of the profile files, named by the profile names.
}
\description{
Helper: Style profiles that can be found
}
