% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_format_time.R
\name{helper_time_format}
\alias{helper_time_format}
\title{Formats time as HH:MM:SS,mmm}
\usage{
helper_time_format(
  t,
  digits = 1,
  addHrsMinSec = FALSE,
  addSec = FALSE,
  format = NULL,
  fps = 25
)
}
\arguments{
\item{t}{Double; time in seconds.}

\item{digits}{Integer; number of digits.}

\item{addHrsMinSec}{Logical; if \code{TRUE} 'hrs' 'min' 'sec' will be used instead of ':'.}

\item{addSec}{Logical; if \code{TRUE} time value in seconds will be shown, too.}

\item{format}{Character or \code{NULL}; a named time format. If set, \code{digits}, \code{addHrsMinSec} and \code{addSec} are ignored. One of \code{"h:mm:ss.s"}, \code{"h:mm:ss.ss"}, \code{"h:mm:ss:ff"} (frames), \code{"mm:ss.ss"} (hours are dropped), \code{"s.s"}, \code{"s.ss"}. Fractions are truncated, not rounded.}

\item{fps}{Numeric; frames per second for \code{"h:mm:ss:ff"}.}
}
\value{
Character string.
}
\description{
Formats time as HH:MM:SS,mmm
}
\examples{
library(act)


helper_time_format(12734.2322345)
helper_time_format(2734.2322345)
helper_time_format(34.2322345)
helper_time_format(0.2322345)

helper_time_format(12734.2322345, addHrsMinSec=TRUE)
helper_time_format(2734.2322345, addHrsMinSec=TRUE)
helper_time_format(34.2322345, addHrsMinSec=TRUE)
helper_time_format(0.2322345, addHrsMinSec=TRUE)

helper_time_format(12734.2322345, digits=3)
helper_time_format(2734.2322345, digits=3)
helper_time_format(34.2322345, digits=3)
helper_time_format(0.2322345, digits=3)

helper_time_format(12734.2322345, addHrsMinSec=TRUE, digits=3)
helper_time_format(2734.2322345, addHrsMinSec=TRUE, digits=3)
helper_time_format(34.2322345, addHrsMinSec=TRUE, digits=3)
helper_time_format(0.2322345, addHrsMinSec=TRUE, digits=3)

helper_time_format(12734.2322345, addHrsMinSec=TRUE, addSec=TRUE)
helper_time_format(2734.2322345, addHrsMinSec=TRUE, addSec=TRUE)
helper_time_format(34.2322345, addHrsMinSec=TRUE, addSec=TRUE)
helper_time_format(0.2322345, addHrsMinSec=TRUE, addSec=TRUE)

helper_time_format(12734.2322345, addHrsMinSec=TRUE, digits=3, addSec=TRUE)
helper_time_format(2734.2322345, addHrsMinSec=TRUE, digits=3, addSec=TRUE)
helper_time_format(34.2322345, addHrsMinSec=TRUE, digits=3, addSec=TRUE)
helper_time_format(0.2322345, addHrsMinSec=TRUE, digits=3, addSec=TRUE)

helper_time_format(83.45, format="h:mm:ss.ss")
helper_time_format(83.45, format="h:mm:ss:ff", fps=25)

}
