% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_format_time.R
\name{helper_time_formats_list}
\alias{helper_time_formats_list}
\title{Helper: Time formats}
\usage{
helper_time_formats_list()
}
\value{
Character vector of format names.
}
\description{
Lists the time formats that \link{helper_time_format} accepts as
\code{format}.
}
\examples{
act::helper_time_formats_list()
}
