% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_transcriptNames_get.R
\name{helper_transcriptNames_get}
\alias{helper_transcriptNames_get}
\title{Helper: Get names of all transcripts in a corpus}
\usage{
helper_transcriptNames_get(x)
}
\arguments{
\item{x}{Corpus object}
}
\value{
List
}
\description{
Gets the names of all transcript objects in a corpus object based from the \code{@name} attribute of each transcript.
}
\examples{
library(act)

act::helper_transcriptNames_get(examplecorpus)

}
