% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_srt.R
\name{import_srt}
\alias{import_srt}
\title{Import a single .srt 'subrib title'  object}
\usage{
import_srt(filePath, transcriptName = NULL, tierName = "subtitle")
}
\arguments{
\item{filePath}{Character string; input path of a single 'subribtitle' .srt file.}

\item{transcriptName}{Character string; name of the transcript.}

\item{tierName}{Character string; name of the imported tier}
}
\value{
Transcript object.
}
\description{
Advice: In most situations it is more convenient to use \code{act::corpus_new}, \code{act::corpus_import} for importing annotation files.
}
\examples{
library(act)

# Path to the .TextGrid file that you want to read
path <- system.file("extdata", "examplecorpus", "GAT", 
					"ARG_I_PAR_Beto.TextGrid", package="act")

# To import a .TextGrid file of your choice:
\dontrun{
	path <- "PATH_TO_AN_EXISTING_TEXTGRID_ON_YOUR_COMPUTER"
}

# Make sure to have rPraat installed before you try the following
\dontrun{
	# Read TextGrid file with rPraat
	rPraat.tg <- rPraat::tg.read(path)

	# Convert to an act transcript
	t <- act::import_rpraat(rPraat.tg)
	
	# Change the name and add it to the examplecorpus
	t@name <-"rpraat"
	newcorpus <- act::transcripts_add(examplecorpus, t)
	
	# Have a look
	newcorpus@transcripts[["rpraat"]]
	
	# Alternatively, you can use the general import function
	t <- act::import(rPraat.tg)
}
}
\seealso{
\link{corpus_import}, \link{corpus_new}, \link{import}, \link{import_eaf}, \link{import_exb}, \code{import_textgrid}, \link{import_rpraat}
}
