% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_info.R
\name{info}
\alias{info}
\title{Information about corpus and transcript objects}
\usage{
info(...)
}
\arguments{
\item{...}{object; either  a corpus or a transcript object.}
}
\value{
List.
}
\description{
Gives detailed information about the contents of a corpus object or a transcript object that is passed as parameter to the function.
In the case that you want to pass a transcript object form a corpus object, make sure that you access the transcript using double [[]] brackets.
}
\details{
To get summarized information about the transcript and corpus objects use \code{act::info_summarized}.
}
\examples{
library(act)

act::info(examplecorpus)

act::info(examplecorpus@transcripts[[1]])


}
\seealso{
\link{info_summarized}
}
