% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_info_summarized.R
\name{info_summarized}
\alias{info_summarized}
\title{Summarized information about corpus and transcript objects}
\usage{
info_summarized(...)
}
\arguments{
\item{...}{object; either  a corpus or a transcript object.}
}
\value{
List.
}
\description{
Gives summarized information about the contents of a corpus object or a transcript object that is passed as parameter to the function.
In the case that you want to pass a transcript object form a corpus object, make sure that you access the transcript using double [[]] brackets.
}
\details{
To get more detailed information about the tiers in a corpus object use \code{act::info}.
}
\examples{
library(act)

act::info_summarized(examplecorpus)

act::info_summarized(examplecorpus@transcripts[[1]])

}
\seealso{
\link{info}
}
