% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output.R
\name{make_output}
\alias{make_output}
\title{Create png with predetermined options}
\usage{
make_output(file)
}
\arguments{
\item{file}{file name output}
}
\value{
open a png file, must be closed afterward
}
\description{
Create png with predetermined options
}
