% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_zip.R
\name{make_zip}
\alias{make_zip}
\title{Create and update zip file}
\usage{
make_zip(zip_path, zip_name = "Backup_Data", zip_files, file_path = NULL)
}
\arguments{
\item{zip_path}{path to zip archive}

\item{zip_name}{name of the archive. Do not add ".zip" at the end}

\item{zip_files}{filenames of files to add in the archive}

\item{file_path}{path to file to archive. default is identical as zip_path.}
}
\value{

}
\description{
Create and update zip file
}
