% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_delete.R
\name{media_delete}
\alias{media_delete}
\title{Delete media files links from transcript objects}
\usage{
media_delete(x, transcriptNames = NULL)
}
\arguments{
\item{x}{Corpus object.}

\item{transcriptNames}{Vector of character strings; Names of the transcripts for which you want to search media files; leave empty if you want to search media for all transcripts in the corpus object.}
}
\value{
Corpus object.
}
\description{
Delete media files links from transcript objects
}
\examples{
library(act)

examplecorpus <- act::media_delete(examplecorpus)

}
\seealso{
\link{media_assign}, \link{media_getPathToExistingFile}
}
