% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_format_names.R
\name{media_format_names}
\alias{media_format_names}
\title{Format media names}
\usage{
media_format_names(
  x,
  recreateNames = TRUE,
  pattern,
  replacement = NULL,
  conditionalPreserve = FALSE
)
}
\arguments{
\item{x}{Corpus object.}

\item{recreateNames}{Logical; if \code{TRUE} the names will be created again from the media file names , if \code{FALSE} the existing values in the names attributes will be used.}

\item{pattern}{Character strings; Regular Expression as search pattern.}

\item{replacement}{Character strings; Replacement pattern.}

\item{conditionalPreserve}{Logical; if \code{TRUE} the names will be set to an empty string "" if there is only one media file of a kind (audio/video) , if \code{FALSE} all names will be preserved.}
}
\value{
Corpus object.
}
\description{
Format the names attribute of the media paths.
Relevant for media cuts with \link{search_cuts}.
Names attribute can be used to differentiate multiple cuts with of the same media type (e.g. mp4 from different cameras).
}
\examples{
library(act)

# Set the folder(s) where your media files are located in the corpus object
# Please be aware that that the example corpus that comes with the package
# does NOT contain media files. Please download the entire example corpus
# with media files if you want to use this function reasonably.
examplecorpus@paths.media.files <- c("", "")

examplecorpus <- act::media_assign(examplecorpus)

}
\seealso{
\link{media_assign}, \link{media_delete}, \link{search_cuts}, \
}
