% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_remove.R
\name{media_remove}
\alias{media_remove}
\title{Remove media files links from transcript objects}
\usage{
media_remove(x = NULL, filterTranscript = "")
}
\arguments{
\item{x}{Corpus object.}

\item{filterTranscript}{Character string; if defined, only transcripts that match this regular expression will be affected.}
}
\value{
Corpus object.
}
\description{
Remove media files links from transcript objects
}
\examples{
examplecorpus <- act::media_remove(examplecorpus)

}
\seealso{
\link{media_assign}, \link{media_getpath}
}
