% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_img.R
\name{merge_img}
\alias{merge_img}
\title{Merge multiple pictures into one}
\usage{
merge_img(img_pattern, path = "~/")
}
\arguments{
\item{img_pattern}{#name pattern shared by the images}

\item{path}{path to pictures}
}
\value{
png
}
\description{
Merge multiple pictures into one
}
