% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max.R
\name{min_max}
\alias{min_max}
\title{Combine min and max functions}
\usage{
min_max(data, which = F)
}
\arguments{
\item{data}{data frame}
}
\value{
c(min, max)
}
\description{
Combine min and max functions
}
\examples{
data("cars")
min_max(cars)
}
