% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_options.R
\name{options_delete}
\alias{options_delete}
\title{Delete all options set by the package from R options}
\usage{
options_delete()
}
\description{
Delete all options set by the package from R options
}
\examples{
library(act)
act::options_delete()
}
