% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_options.R
\name{options_remove}
\alias{options_remove}
\title{Remove all options set by the package from R options}
\usage{
options_remove()
}
\description{
Remove all options set by the package from R options
}
\examples{
library(act)
act::options_remove()
}
