% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_column.R
\name{order_column}
\alias{order_column}
\title{Put the selected columns at the beginning of the dataset.}
\usage{
order_column(data, col_names)
}
\arguments{
\item{data}{data set}

\item{col_names}{columns to put first}
}
\value{

}
\description{
Put the selected columns at the beginning of the dataset.
}
\examples{
data("iris")
order_column(iris, "Species")
}
