% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_number.R
\name{plot_number}
\alias{plot_number}
\title{Return the number of the current plot.}
\usage{
plot_number(out = "pos")
}
\arguments{
\item{out}{"pos" give plot number. "x" give column number of the current plot, "y" give row number of the current plot}
}
\value{
Plot number in current window.
}
\description{
Return the number of the current plot.
}
\examples{
plot_number()
}
