% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_clean.R
\name{read_csv_clean}
\alias{read_csv_clean}
\title{Import a csv file while removing duplicate row and keeping numerical col as numeric.}
\usage{
read_csv_clean(df_path, dec = ".")
}
\arguments{
\item{df_path}{The name of the file which the data are to be read from.}

\item{dec}{The character used in the file for decimal points.}
}
\value{
data frame
}
\description{
Import a csv file while removing duplicate row and keeping numerical col as numeric.
}
