% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression}
\alias{regression}
\title{Compute and draw the Linear Regression on a plot}
\usage{
regression(
  df,
  x,
  y,
  type,
  intercept = NULL,
  output = 1,
  filename = paste0("regression_", Sys.Date(), ".csv"),
  print.ann = T,
  ann.pos = "right",
  col = "grey50"
)
}
\arguments{
\item{df}{data frame}

\item{x}{x parameter}

\item{y}{y parameter}

\item{type}{If "nls" or "lm" return a second order regression. If numeric, return a regression from the specified order, using the lm function ("lm" return the same results than 1).}

\item{intercept}{numeric, force the intercept value}

\item{output}{1 draws a regression onto the current plot. 2 writes the regression parameters in a data frame. 3 do both.}

\item{filename}{output filename. csv format by default}

\item{print.ann}{logical. add regression equation and r square on the plot}

\item{ann.pos}{equation position}

\item{col}{line color}
}
\value{
draw a curve on a pre-existing plot
}
\description{
Compute and draw the Linear Regression on a plot
}
