% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsm_plot.R
\name{rsm_plot}
\alias{rsm_plot}
\title{Response Surface Methods plot
Build a second order model from the variables and draw a contour plot and a perspective plot.}
\usage{
rsm_plot(
  data,
  x,
  y,
  z,
  method = c("contour", "persp"),
  col_pal = colorRampPalette(c("#660000", "#ff0000"))(8)
)
}
\arguments{
\item{data}{Data Frame}

\item{x}{First factor}

\item{y}{Second factor}

\item{z}{Variable}

\item{method}{"contour" and / or "persp", Default is both.}

\item{col_pal}{color palette}
}
\value{
RSM plot
}
\description{
Response Surface Methods plot
Build a second order model from the variables and draw a contour plot and a perspective plot.
}
\examples{
data("iris")
par(mfrow = c(1, 2))
rsm_plot(iris, "Sepal.Length", "Sepal.Width", "Petal.Length")
}
