% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_cuts_srt.R
\name{search_cuts_srt}
\alias{search_cuts_srt}
\title{Create .srt subtitles for all search results}
\usage{
search_cuts_srt(
  x,
  s,
  cutSpanBeforesec = NULL,
  cutSpanAftersec = NULL,
  folderOutput = NULL,
  speakerShow = TRUE,
  speakerWidth = 3,
  speakerEnding = ":"
)
}
\arguments{
\item{x}{Corpus object.}

\item{s}{Search object.}

\item{cutSpanBeforesec}{Double; Start the cut some seconds before the hit to include some context; the default NULL will take the value as set in @cuts.span.beforesec of the search object.}

\item{cutSpanAftersec}{Double; End the cut some seconds before the hit to include some context; the default NULL will take the value as set in @cuts.span.beforesec of the search object.}

\item{folderOutput}{Character string; path to an existing folder for file export. If \code{NULL} (default), no files will be written to disk; the srt subtitles are inserted only into \code{s@results}. If set, .srt files will be written to disk in addition.}

\item{speakerShow}{Logical; if \code{TRUE} name of speaker will be shown before the content of the annotation.}

\item{speakerWidth}{Integer; width of speaker abbreviation, -1 for full name without shortening.}

\item{speakerEnding}{Character string; string that is added at the end of the speaker name.}
}
\value{
Search object;
}
\description{
Subtitles in 'Subrib Title' .srt format will be created for each search result.
The subtitles will be inserted into the column defined in \code{s@cuts.column.srt}.
}
\details{
\emph{Span} \cr
If you want to extend the cut before or after each search result, you can modify \code{@cuts.span.beforesec} and \code{@cuts.span.aftersec} in your search object.
}
\examples{
library(act)

# Search
mysearch <- act::search_new(examplecorpus, pattern="yo")

# Create srt subtitles for all search results
test <- act::search_cuts_srt (x=examplecorpus, s=mysearch)

# Display srt subtitle of first three results
cat(test@results[1:3, mysearch@cuts.column.srt])

# Create srt subtitle including 1 sec before and 5 sec after
mysearch@cuts.span.beforesec = 1
mysearch@cuts.span.aftersec = 5
test <- act::search_cuts_srt (x=examplecorpus,
								s=mysearch)

# Display srt subtitle of first  results
cat(test@results[1,mysearch@cuts.column.srt])
}
