% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_openresult_inpraat.R
\name{search_openresult_inpraat}
\alias{search_openresult_inpraat}
\title{Open a search result in 'Praat'}
\usage{
search_openresult_inpraat(
  x,
  s,
  resultid,
  play = TRUE,
  close = FALSE,
  filterMediaFile = c("(?i).*\\\\.(aiff|aif|wav)", "(?i).*\\\\.mp3"),
  delay = 0.5
)
}
\arguments{
\item{x}{Corpus object.}

\item{s}{Search object.}

\item{resultid}{Integer; Number of the search result (row in the data frame \code{s@results}) to be played.}

\item{play}{Logical; If \code{TRUE} selection will be played.}

\item{close}{Logical; If \code{TRUE} TextGrid editor will be closed after playing (Currently non functional!)}

\item{filterMediaFile}{Vector of character strings; Each element of the vector is a regular expression. Expressions will be checked consecutively. The first match with an existing media file will be used for playing. The default checking order is uncompressed audio > compressed audio.}

\item{delay}{Double; Time in seconds before the section will be opened in Praat. This is useful if Praat opens but the section does not. In that case increase the delay.}
}
\description{
The function remote controls 'Praat' with a 'Praat' script.
It opens a search result in the 'Praat' TextGrid Editor.
}
\details{
'Praat' receives the script through its own command line option \code{--send}
(from 'Praat' 6.1 on): set the path to 'Praat' with 'options(act.path.praat = ...)'.
If 'Praat' is not running, it is started. Only if the path to 'Praat' is not set,
'sendpraat' is used ('options(act.path.sendpraat = ...)', see \code{vignette("installation-sendpraat")}).
'Praat' 7 runs a 'sendpraat' message late, so the path to 'Praat' is the better choice.
}
\examples{
library(act)

mysearch <- act::search_new(x=examplecorpus, pattern = "pero")

# You can only use this functions if you have installed 'Praat'
# and located it properly in the package options.
\dontrun{
act::search_openresult_inpraat(x=examplecorpus, s=mysearch, resultid=1, TRUE, TRUE)
}
}
\seealso{
\code{vignette("install_sendpraat", package = "act")}
}
