% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_searchandopen_inpraat.R
\name{search_searchandopen_inpraat}
\alias{search_searchandopen_inpraat}
\title{Search corpus and open first result in Praat}
\usage{
search_searchandopen_inpraat(x, pattern)
}
\arguments{
\item{x}{Corpus object.}

\item{pattern}{Character string; search pattern as regular expression.}
}
\description{
The function remote controls 'Praat' by using 'sendpraat' and a 'Praat' script.
It first searches your corpus object and uses the first search hit.
The corresponding TextGrid will be opened in the 'Praat' TextGrid Editor and the search hit will be displayed.
}
\details{
To make this function work you need to do two things first:
\itemize{
\item Install 'sendpraat' on your computer. To do so  follow the  instructions in the vignette 'installation-sendpraat'. Show the vignette with \code{vignette("installation-sendpraat")}.
\item Set the path to the 'sendpraat' executable correctly by using 'options(act.path.sendpraat = ...)'.
}
}
\examples{
library(act)

# You can only use this functions if you have installed
# and located the 'sendpraat' executable properly in the package options.
\dontrun{
act::search_searchandopen_inpraat(x=examplecorpus, "pero")
}



}
\seealso{
\code{vignette("install_sendpraat", package = "act")}
}
