% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_stills.R
\name{search_stills}
\alias{search_stills}
\title{Search stills for a search}
\usage{
search_stills(
  x,
  s,
  pattern,
  searchMode = c("content", "fulltext", "fulltext.byTime", "fulltext.byTier"),
  searchNormalized = FALSE,
  filterTierNames = NULL,
  filterTierIncludeRegex = "still",
  filterTierExcludeRegex = NULL,
  resultids = NULL,
  stillsFolder = "stills"
)
}
\arguments{
\item{x}{Corpus object; basis in which will be searched.}

\item{s}{Search object.}

\item{pattern}{Character string; search pattern as regular expression.}

\item{searchMode}{Character string; takes the following values: \code{content}, \code{fulltext} (=default, includes both full text modes), \code{fulltext.byTime}, \code{fulltext.byTier}.}

\item{searchNormalized}{Logical; if \code{TRUE} function will search in the normalized content, if \code{FALSE} function will search in the original content.}

\item{filterTierNames}{Vector of character strings; names of tiers to be included.}

\item{filterTierIncludeRegex}{Character string; as regular expression, exclude some tiers from search matching the expression.}

\item{resultids}{Vector of Integer; By default all results in the search object will be processed.  if \code{resultids} is set, still will only be searched for the results specified.}

\item{stillsFolder}{Character string; name of the sub folder where to store the stills. folder will be created recursively}
}
\value{
Search object.
}
\description{
Only \code{x}, \code{s} and \code{pattern} are obligatory.
The other arguments can be left to their default values.
}
\examples{
library(act)

# Search for the 1. Person Singular Pronoun in Spanish.
mysearch <- act::search_new(examplecorpus, pattern= "yo")
mysearch
# Search in normalized content vs. original content
mysearch.norm  <- act::search_new(examplecorpus, pattern="yo", searchNormalized=TRUE)
mysearch.org   <- act::search_new(examplecorpus, pattern="yo", searchNormalized=FALSE)
mysearch.norm@results.nr
mysearch.org@results.nr

# The difference is because during normalization capital letters will be converted
# to small letters. One annotation in the example corpus contains a "yo" with a
# capital letter:
mysearch <- act::search_new(examplecorpus, pattern="yO", searchNormalized=FALSE)
mysearch@results$hit

# Search in full text vs. original content.
# Full text search will find matches across annotations.
# Let's define a regular expression with a certain span.
# Search for the word "no" 'no' followed by a "pero" 'but'
# in a distance ranging from 1 to 20 characters.
myRegEx <- "\\\\bno\\\\b.{1,20}pero"
mysearch <- act::search_new(examplecorpus, pattern=myRegEx, searchMode="fulltext")
mysearch
mysearch@results$hit

}
\seealso{
\link{search_cuts_media}, \link{search_new}
}
