% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_transcript_content.R
\name{search_transcript_content}
\alias{search_transcript_content}
\title{Search in original content of a single transcript}
\usage{
search_transcript_content(t, s)
}
\arguments{
\item{t}{Transcript object; transcript to search in.}

\item{s}{Search object.}
}
\value{
Data.frame; data frame with search results.
 
# @example inst/examples/search_transcript_content.R
}
\description{
Search in original content of a single transcript
}
