% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shades_of_grey.R
\name{shades_of_grey}
\alias{shades_of_grey}
\title{ShadesOfGrey
Color palette with shade from white to black}
\usage{
shades_of_grey(n)
}
\arguments{
\item{n}{Number of colors to return. Max is 100.}
}
\value{
Hex color value
}
\description{
ShadesOfGrey
Color palette with shade from white to black
}
\examples{
shades_of_grey(3)
}
