% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_analysis.R
\name{stat_analysis}
\alias{stat_analysis}
\title{Make multiple statistical analysis on a data set
lm(Y ~ factors, data)}
\usage{
stat_analysis(
  data,
  Y,
  factors,
  sep = NULL,
  test = c(1:4),
  mfrow = NULL,
  SavePath = NULL,
  l = F,
  eb = F
)
}
\arguments{
\item{data}{data set}

\item{Y}{Variable}

\item{factors}{Factors}

\item{sep}{Multiple loop condition}

\item{test}{Choose the test to effectuate. 1 = ANOVA, 2 = TukeyHSD, 3 = BargroupHSD, 4 = ANOVA2. Default is c(1:4).}

\item{mfrow}{plot number}

\item{SavePath}{If fill, save all the analysis into the indicated folder}

\item{l}{Add letters to plot if TRUE.}

\item{eb}{Add errorbars to plot if TRUE.}
}
\value{
ANOVA, TukeyHSD, Bargroup HSD,
}
\description{
Make multiple statistical analysis on a data set
lm(Y ~ factors, data)
}
\examples{
data("CO2")
stat_analysis(CO2, "uptake", c("Treatment", "Type"))
}
