% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiers_all.R
\name{tiers_all}
\alias{tiers_all}
\title{All tiers in a corpus}
\usage{
tiers_all(x, compact = TRUE)
}
\arguments{
\item{x}{Corpus object.}

\item{compact}{Logical; if \code{TRUE} a condensed overview will be returned,  if \code{FALSE} a detailed overview will be returned.}
}
\value{
Data frame
}
\description{
Merges tiers from all transcripts in a corpus object into a data frame.
}
\examples{

library(act)

#Get data frame with all tiers
alltiers <- act::tiers_all(examplecorpus)
alltiers

#Get data frame with a simplified version
alltiers <- act::tiers_all(examplecorpus, compact=TRUE)
alltiers

}
