% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts_cure_single.R
\name{transcripts_cure_single}
\alias{transcripts_cure_single}
\title{Cure a single transcript}
\usage{
transcripts_cure_single(
  t,
  annotationsTimesReversed = TRUE,
  annotationsOverlap = TRUE,
  annotationsTimesBelowZero = TRUE,
  transcriptLengthZero = TRUE,
  annotationsZeroDuration = TRUE,
  tiersMissing = TRUE,
  warning = FALSE
)
}
\arguments{
\item{t}{Transcript object.}

\item{annotationsTimesReversed}{Logical; If \code{TRUE} annotations with reversed times will be deleted}

\item{annotationsOverlap}{Logical; If \code{TRUE} overlapping annotations will be corrected.}

\item{annotationsTimesBelowZero}{Logical; If \code{TRUE} annotations before 0 sec will be corrected.}

\item{transcriptLengthZero}{Logical; If \code{TRUE} a transcript with \code{@length.sec <= 0} will be corrected to the largest annotation endsec.}

\item{annotationsZeroDuration}{Logical; If \code{TRUE} zero-duration annotations in IntervalTiers will be merged/extended/deleted.}

\item{tiersMissing}{Logical; If \code{TRUE} tiers missing in \code{@tiers} slot of the transcript object will be added.}

\item{warning}{Logical; If \code{TRUE} a warning notice will be shown upon correction.}
}
\value{
Transcript object;
}
\description{
Transcript object may contain errors, e.g. because of defect annotation input files or user modifications.
This function may cure some of these errors.
\itemize{
\item Annotations with reversed times: annotations with \code{endsec} lower than \code{startsec} will be deleted.
\item Overlapping annotations: earlier annotations will end where the next annotation starts.
\item Annotations below 0 sec: Annotations that are starting and ending before 0 sec will be deleted; Annotations starting before but ending after 0 sec will be truncated.
\item Transcript length zero: If \code{@length.sec} is missing or \code{<= 0} it will be set to the largest \code{endsec} of any annotation (or a small default).
\item Zero-duration annotations: In IntervalTiers, annotations with \code{startsec == endsec} will be cured. Multiple zero-duration annotations at the same time point are merged by concatenating their content with a single space. The resulting annotation is then extended to the start of the next annotation (or to \code{@length.sec} if it is the last one). If no room is available (next annotation starts at the same time), the annotation is deleted.
\item Missing tiers: Tiers that are present in the annotations but missing in the list of tiers in \code{@tiers} of the transcript object will be added.
}
}
\examples{
library(act)

# --- annotationsTimesReversed: will be deleted
# get example transcript and reverse the times of an annotation
t <- examplecorpus@transcripts[[1]]
t@annotations$startsec[1] <- 20
t@annotations$endsec[1] <- 10
t2 <- act::transcripts_cure_single(t)
tail(t2@history, n=1)


# --- annotationsTimesBelowZero: will be deleted or start at 0 sec
t <- examplecorpus@transcripts[[1]]
t@annotations$startsec[1] <- -2
t@annotations$endsec[1]   <- -1
t2 <- act::transcripts_cure_single(t)
tail(t2@history, n=1)

t <- examplecorpus@transcripts[[1]]
t@annotations$startsec[2] <- -5
t2 <- act::transcripts_cure_single(t)
tail(t2@history, n=1)


# --- annotationsOverlap: will end where the next starts
t<- examplecorpus@transcripts[[1]]
t@annotations <- t@annotations[order(t@annotations$tierName, t@annotations$startsec), ]
t@annotations$endsec[1] <- 8
t2 <- act::transcripts_cure_single(t)
tail(t2@history, n=1)


# --- tiersMissing: will be added to @tiers in transcript object
t<- examplecorpus@transcripts[[1]]
t@annotations <- t@annotations[order(t@annotations$tierName, t@annotations$startsec), ]
t@annotations$tierName[1] <- "NEW"
t2 <- act::transcripts_cure_single(t)
tail(t2@history, n=1)
t2@tiers
# compare with original tiers
t@tiers


# --- several things at once
t<- examplecorpus@transcripts[[1]]
t@annotations <- t@annotations[order(t@annotations$tierName, t@annotations$startsec), ]
# annotation completely below 0 sec
t@annotations$startsec[1] <- -6
t@annotations$endsec[1]   <- -5
# annotation starts before but ends after 0 sec
t@annotations$startsec[2] <- -3
# annotation with reversed times
t@annotations$startsec[3] <- 6.9
t@annotations$endsec[3]   <- -6.8
# annotation overlaps with next annotation
t@annotations$endsec[6]   <- 9
# new tier, missing tier list
t@annotations$tierName[8] <- "NEW"
t2 <- act::transcripts_cure_single(t, warning=TRUE)
tail(t2@history, n=1)


examplecorpus@transcripts[[1]]@history
}
\seealso{
\link{transcripts_cure}
}
