% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts_update_normalization.R
\name{transcripts_update_normalization}
\alias{transcripts_update_normalization}
\title{Normalize transcriptions}
\usage{
transcripts_update_normalization(
  x,
  transcriptNames = NULL,
  forceUpdate = FALSE
)
}
\arguments{
\item{x}{Corpus object.}

\item{transcriptNames}{Vector of character strings; Names of the transcripts for which you want to search media files; leave empty if you want to search media for all transcripts in the corpus object.}

\item{forceUpdate}{Logical; If \code{TRUE} transcripts will be normalized in any case, if \code{FALSE} transcripts will be only normalized if there was a modification to the transcript since the last normalization.}
}
\description{
Normalizes the contents of transcriptions in a corpus object using a normalization matrix.
Function returns a corpus object with normalized transcription and updates the original corpus object passed as argument to x.
}
\details{
If the normalization matrix is missing or empty (0 rows), a warning is issued
and the corpus object is returned unchanged.
}
\examples{
library(act)

examplecorpus <- act::transcripts_update_normalization(x=examplecorpus)

}
