% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc2.R
\name{trunc2}
\alias{trunc2}
\title{Truncate number after n decimal}
\usage{
trunc2(x, n)
}
\arguments{
\item{x}{value to round}

\item{n}{decimal number}
}
\value{
numeric
}
\description{
Truncate number after n decimal
}
\examples{
trunc2(1.23456, 2)
}
