% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_table.R
\name{unique_table}
\alias{unique_table}
\title{Create an ordered dataframe with all unique combination of a data set}
\usage{
unique_table(data, factors)
}
\arguments{
\item{data}{data frame}

\item{factors}{Either column numbers or column names}
}
\value{
Data frame with all combinations
}
\description{
Create an ordered dataframe with all unique combination of a data set
}
\examples{
data("CO2")
unique_table(CO2, c(1:3))
}
