% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_docx.R
\name{export_docx}
\alias{export_docx}
\title{Export print transcript in .docx format}
\usage{
export_docx(
  t,
  l = NULL,
  pathOutput = NULL,
  filterTierNames = NULL,
  filterSectionStartsec = NULL,
  filterSectionEndsec = NULL,
  insertArrowAnnotationID = "",
  headerPreface = NULL,
  headerTitle = NULL,
  headerSubtitle = NULL,
  headerDescription = NULL,
  headerInsertSource = TRUE
)
}
\arguments{
\item{t}{Transcript object.}

\item{l}{Layout object.}

\item{pathOutput}{Character string; path where to save the transcript.}

\item{filterTierNames}{Vector of character strings; names of tiers to be included. If left unspecified, all tiers will be exported.}

\item{filterSectionStartsec}{Double; start of selection in seconds.}

\item{filterSectionEndsec}{Double; end of selection in seconds.}

\item{insertArrowAnnotationID}{Integer; ID of the annotation in front of which the arrow will be placed.}

\item{headerPreface}{Character string; text used as preface before title.}

\item{headerTitle}{Character string; text used as title.}

\item{headerSubtitle}{Character string; text  used as sub title.}

\item{headerDescription}{Character string; text used as description after sub title.}

\item{headerInsertSource}{Logical; if \code{TRUE} standard information about the source and location of the sequence will be inserted after the heading.}
}
\value{
Officer doc; transcript as object from library officer.
}
\description{
LAYOUT
If you want to modify the layout of the print transcripts, create a new layout object with \code{mylayout <- methods::new("layout")}, modify the settings and pass it as argument \code{l}.
Using the layout object you may
\itemize{
\item Adjust with, abbreviation of speakers, etc.
\item set filters to include/exclude tiers matching regular expressions.
\item assign template files for .docx formatting using format templates
}
}
\details{
FORMATING
\itemize{
\item adjust the the defaults format templates in the default .docx template.
\item define further templates and add them to a styles matrix.
The paths to both files need to be set in your l layout object. Please check the slots l@docx.template.path and l@docx.styles.base.
You can see the structure of the default styles matrix in each new layout object in l@docx.styles.base. Use l@docx.styles.base <- act::export_styles_base_load(...) to assign a custom styles matrix.
The default format templates are
}
\itemize{
\item Header:
}
\itemize{
\item header.preface (formats: s@results$header.description)
\item header.title (formats: s@results$header.description)
\item header.subtitle (formats: s@results$header.description)
\item header.description (formats: s@results$header.description)
}
\itemize{
\item Transcript body
}
\itemize{
\item transcript.default (formats: any annotation in "t@annotations"
}
}
\examples{
library(act)

# Get a transcript
t <- examplecorpus@transcripts[[1]]

# Create print transcript
printtranscript <- act::export_docx (t=t)

# Display on screen
printtranscript

}
\seealso{
\link{corpus_export}, \link{export_eaf}, \link{export_exb}, \link{export_rpraat}, \link{export_srt}, \link{export_textgrid}, \code{vignette("export_docx_styles", package = "act")}
}
