% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_txt.R
\name{export_txt}
\alias{export_txt}
\alias{export_printtranscript}
\title{Export print transcript in .txt format}
\usage{
export_txt(
  t,
  l = NULL,
  pathOutput = NULL,
  filterTierNames = NULL,
  filterSectionStartsec = NULL,
  filterSectionEndsec = NULL,
  insertArrowAnnotationID = "",
  headerPreface = NULL,
  headerTitle = NULL,
  headerSubtitle = NULL,
  headerDescription = NULL,
  headerInsertSource = TRUE,
  collapse = TRUE
)

export_printtranscript(...)
}
\arguments{
\item{t}{Transcript object.}

\item{l}{Layout object.}

\item{pathOutput}{Character string; path where to save the transcript.}

\item{filterTierNames}{Vector of character strings; names of tiers to be included. If left unspecified, all tiers will be exported.}

\item{filterSectionStartsec}{Double; start of selection in seconds.}

\item{filterSectionEndsec}{Double; end of selection in seconds.}

\item{insertArrowAnnotationID}{Integer; ID of the annotation in front of which the arrow will be placed.}

\item{headerPreface}{Character string; text used as preface before title.}

\item{headerTitle}{Character string; text used as title.}

\item{headerSubtitle}{Character string; text  used as sub title.}

\item{headerDescription}{Character string; text used as description after sub title.}

\item{headerInsertSource}{Logical; if \code{TRUE} standard information about the source and location of the sequence will be inserted after the heading.}

\item{collapse}{Logical; if \code{FALSE} a vector will be created, each element corresponding to one annotation. if \code{TRUE} a single string will be created, collapsed by linebreaks \\n.}

\item{...}{Arguments passed to \code{export_txt()}.}
}
\value{
Character string; transcript as text.
}
\description{
If you want to modify the layout of the print transcripts, create a new layout object with \code{mylayout <- methods::new("layout")}, modify the settings and pass it as argument \code{l}.
In the layout object you may also set additional filters to include/exclude tiers matching regular expressions.
}
\examples{
library(act)

# Get a transcript
t <- examplecorpus@transcripts[[1]]

# Create print transcript
printtranscript <- act::export_txt (t=t)

# Display on screen
cat(printtranscript)

}
\seealso{
\link{corpus_export}, \link{export_eaf}, \link{export_exb}, \link{export_rpraat}, \link{export_srt}, \link{export_textgrid}, \link{export_docx}

\code{\link{export_txt}}
}
