% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_tiers_filter_create.R
\name{helper_tiers_filter_create}
\alias{helper_tiers_filter_create}
\title{Helper: Create filter for tier names}
\usage{
helper_tiers_filter_create(
  tierNames,
  filterTierIncludeRegEx = "",
  filterTierExcludeRegEx = ""
)
}
\arguments{
\item{tierNames}{Vector of character strings; names of the tiers.}

\item{filterTierIncludeRegEx}{Character string; as regular expression, tiers matching the expression will be included in the print transcript.}

\item{filterTierExcludeRegEx}{Character string; as regular expression, tiers matching the expression will be excluded from the print transcript.}
}
\value{
Vector of character strings; names of tiers
}
\description{
Creates a tier name filter based on a vector of character strings and the values 'filterTierIncludeRegEx' and 'filterTierExcludeRegEx' in a layout object.
}
\examples{
library(act)

# --- Create a tier table from scratch
tierTable <- act::helper_tiers_new_table(c("a","b","c", "d"), 
c("IntervalTier", "TextTier","IntervalTier","TextTier"))
tierTable
}
