% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_layout.R
\docType{class}
\name{layout-class}
\alias{layout-class}
\title{Layout object, defining the layout of print transcripts}
\description{
You can create an new layout object with \code{methods::new("layout")}.
This will give you a new layout object with the default settings uses by act.
If you want to modify the layout of the print transcripts, create a new layout object with \code{mylayout <- methods::new("layout")}, modify the values in the \code{@slots} and pass it as argument \code{l} to the respective functions.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character string; Name of the layout.}

\item{\code{filter.tier.includeRegEx}}{Character string; as regular expression, tiers matching the expression will be included in the print transcript.}

\item{\code{filter.tier.excludeRegEx}}{Character string; as regular expression, tiers matching the expression will be excluded from the print transcript.}

\item{\code{transcript.width}}{Integer; width of transcript, -1 for no line wrapping.}

\item{\code{speaker.regex}}{Character string; Regular expression to extract speaker abbreviation from tier name}

\item{\code{speaker.width}}{Integer; maximum width of speaker abbreviation, -1 for full name without shortening.}

\item{\code{speaker.ending}}{Character string; string that is added at the end of the speaker name.}

\item{\code{spacesbefore}}{Integer; number of spaces inserted before line number.}

\item{\code{brackets.align}}{Logical; if \code{TRUE} act will try to align brackets [] for parallel speaking (Attention: experimental function; results may not satisfy).}

\item{\code{header.insert}}{Logical; if \code{TRUE} a transcript header is inserted.}

\item{\code{arrow.insert}}{Logical; is only used when transcripts are made based on a search results; if \code{TRUE} an arrow will be inserted, highlighting the transcript line containing the search hit.}

\item{\code{arrow.shape}}{Character string; shape of the arrow.}

\item{\code{docx.template.path}}{Character string;}

\item{\code{docx.styles.base}}{Data.frame; Matrix with mappings of act style names to DOCX template paragraph styles. To change the styles matrix use \code{l@docx.styles.base <- act::export_styles_base_load(path="...")}}

\item{\code{docx.styles.user}}{Data.frame; Matrix with user-defined tier-specific formatting rules (regex-based). Empty data.frame means all tiers use layout defaults.}
}}

\seealso{
\link{matrix_load}
}
