% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_bar.R
\name{progress_bar}
\alias{progress_bar}
\title{create a basic progression bar
May lag if total if really high.}
\usage{
progress_bar(i, total)
}
\arguments{
\item{i}{iteration}

\item{total}{total number of iteration}
}
\value{
progress bar in console
}
\description{
create a basic progression bar
May lag if total if really high.
}
\examples{
for (i in 1:10) {Sys.sleep(0.1); progress_bar(i, 10)}
}
