% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_cuts.R
\name{search_cuts}
\alias{search_cuts}
\title{Create print transcripts, media cutlists and srt subtitles for all search results}
\usage{
search_cuts(
  x,
  s,
  cutSpanBeforesec = NULL,
  cutSpanAftersec = NULL,
  l = NULL,
  outputFolder = NULL
)
}
\arguments{
\item{x}{Corpus object.}

\item{s}{Search object.}

\item{cutSpanBeforesec}{Double; Start the cut some seconds before the hit to include some context; the default NULL will take the value as set in @cuts.span.beforesec of the search object.}

\item{cutSpanAftersec}{Double; End the cut some seconds before the hit to include some context; the default NULL will take the value as set in @cuts.span.beforesec of the search object.}

\item{l}{Layout object.}

\item{outputFolder}{Character string; if parameter is not set, the print transcripts will only be inserted in \code{s@results}; if the path to a existing folder is given transcripts will be saved in '.txt' format.}
}
\value{
Search object;
}
\description{
This function will call the following functions:
- \code{act_cuts_printtranscript} to create print transcripts,
- \code{act::cuts_media} to create FFmpeg cutlist to make media snippets,
#' - \code{act::search_cuts_srt()} to create .srt subtitles,
for all  search results.
}
\details{
If you want to extend the cut before or after each search result, you can modify \code{@cuts.span.beforesec} and \code{@cuts.span.aftersec} in your search object.
If you want to modify the layout of the print transcripts, create a new layout object with \code{mylayout <- methods::new("layout")}, modify the settings and pass it as argument \code{l}.

For a detailed description and more examples please refer to the documentation of the infidival functions.
They also offer some more parameters than this functions. If you want to use those, call the functions individually.
}
\examples{
library(act)

# Search
mysearch <- act::search_new(examplecorpus, pattern="yo")

# Create print transcripts, media cutlists and .srt subtitles 
# for all search results
test <- act::search_cuts(x=examplecorpus, s=mysearch)

# Display all print transcripts on screen from @cuts.printtranscripts
cat(test@cuts.printtranscripts)

}
