% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_openresult_inpraat.R
\name{search_openresult_inpraat}
\alias{search_openresult_inpraat}
\title{Open a search result in 'Praat'}
\usage{
search_openresult_inpraat(
  x,
  s,
  resultid,
  play = TRUE,
  close = FALSE,
  filterMediaFile = c("(?i).*\\\\.(aiff|aif|wav)", "(?i).*\\\\.mp3"),
  delay = 0.5
)
}
\arguments{
\item{x}{Corpus object.}

\item{s}{Search object.}

\item{resultid}{Integer; Number of the search result (row in the data frame \code{s@results}) to be played.}

\item{play}{Logical; If \code{TRUE} selection will be played.}

\item{close}{Logical; If \code{TRUE} TextGrid editor will be closed after playing (Currently non functional!)}

\item{filterMediaFile}{Vector of character strings; Each element of the vector is a regular expression. Expressions will be checked consecutively. The first match with an existing media file will be used for playing. The default checking order is uncompressed audio > compressed audio.}

\item{delay}{Double; Time in seconds before the section will be opened in Praat. This is useful if Praat opens but the section does not. In that case increase the delay.}
}
\description{
The function remote controls 'Praat' by using 'sendpraat' and a 'Praat' script.
It opens a search result in the 'Praat' TextGrid Editor.
}
\details{
To make this function work you need to do two things first:
\itemize{
\item Install 'sendpraat' on your computer. To do so  follow the  instructions in the vignette 'installation-sendpraat'. Show the vignette with \code{vignette("installation-sendpraat")}.
\item Set the path to the 'sendpraat' executable correctly by using 'options(act.path.sendpraat = ...)'.
}
}
\examples{
library(act)

mysearch <- act::search_new(x=examplecorpus, pattern = "pero")

# You can only use this functions if you have installed and 
# located the 'sendpraat' executable properly in the package options.
\dontrun{
act::search_openresult_inpraat(x=examplecorpus, s=mysearch, resultid=1, TRUE, TRUE)
}
}
\seealso{
\code{vignette("install_sendpraat", package = "act")}
}
