% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_transcript_fulltext.R
\name{search_transcript_fulltext}
\alias{search_transcript_fulltext}
\title{Search in full text of a single transcript}
\usage{
search_transcript_fulltext(t, s)
}
\arguments{
\item{t}{Transcript object; transcript to search in.}

\item{s}{Search object.}
}
\value{
Data.frame; data frame with search results.

  
# @example inst/examples/search_transcript_fulltext.R
}
\description{
Search in full text of a single transcript
}
