% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts_openin_elan.R
\name{transcripts_openin_elan}
\alias{transcripts_openin_elan}
\title{Open transcript in 'ELAN'}
\usage{
transcripts_openin_elan(
  t,
  openOriginal = FALSE,
  filePathOut = NULL,
  startSec = NULL,
  endSec = NULL,
  tierName = NULL
)
}
\arguments{
\item{t}{Transcript object.}

\item{openOriginal}{Logical; if \code{TRUE} the function will check if the original annotation file was an .eaf file and if it still exists in the original location. If so, the function will not create a temporary .eaf file but open the original file. Warning: The original .pfsx file (if it exists) will be overwritten.}

\item{filePathOut}{Character string, optional. If original .eaf is not found, a .eaf file will be created at this location. If \code{NULL} a temporary .eaf file will be created.}

\item{startSec}{Double, Optional, start of selection en ELAN}

\item{endSec}{Double, Optional, end of selection en ELAN}

\item{tierName}{Character string, Optional, Name of tier to select in ELAN grid view.}
}
\description{
The function opens a transcript in ELAN-
It can open the an original .eaf file (if it exists) or create an .eaf file.
A .psfx file will be created id startSec and endSec are defined.
}
\details{
To make this function work you need to have 'ELAN' installed on your computer and tell the act package where ELAN is located.
Therefore you need to set the path to the ELAN executable in the option 'act.path.elan' using \code{options(act.path.elan='PATHTOYOURELANEXECUTABLE')}.

WARNING: This function will overwrite existing .psfx files.

Credits: Thanks to Han Sloetjes for feedback on the structure of the temporary .pfsx files. He actually made the code work.
}
\examples{
library(act)

mysearch <- act::search_new(x=examplecorpus, pattern = "yo")

# You can only use this function if you have installed ELAN on our computer.
\dontrun{
options(act.path.elan='PATHTOYOURELANEXECUTABLE')
act::search_openresult_inelan(x=examplecorpus, s=mysearch, resultid=1, TRUE)
}

}
